Include CMSIS/STM32L4xx core includes from Arduino.h

+++ cores/stm32l4/stm32l4_wiring.h
@@ -54,7 +54,7 @@ static inline uint32_t millis(void)
     return armv7m_systick_millis();
 }
 
-static inline uint64_t micros(void) 
+static inline uint32_t micros(void) 
 {
     return armv7m_systick_micros();
 }
